requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
librosa==0.10.1
pydub==0.25.1
scikit-learn==1.4.0
numpy==1.26.3
python-multipart==0.0.6
soundfile==0.12.1
python-dotenv==1.0.0
